refactor!: move compound layers to the ops module
