+++ src/s2vit/__init__.py
@@ -1,3 +1,3 @@
-from . import nn, transformer, utils
+from . import nn, ops, transformer, utils
 from ._version import __version__
 from .transformer import S2ViT
