+++ src/s2vit/transformer.py
@@ -4,80 +4,19 @@ from collections.abc import Callable, Sequence
 from typing import Any, Literal
 
 import torch
-from einops import pack, rearrange, unpack
 from einops.layers.torch import Reduce
 from torch import nn
-from torch.nn import functional as F
 from torchvision.ops import DropBlock2d
 
-from .nn import (
+from .nn import LayerNormNoBias, LayerNormNoBias2d
+from .ops import (
     PEG,
-    LayerNormNoBias,
-    LayerNormNoBias2d,
+    FusedGWAttentionFF,
     ParallelFF,
     ParallelGWAttention,
     PatchEmbedding,
     Shift2d,
-    StarReLU,
 )
-
-
-class FusedGWAttentionFF(nn.Module):
-    def __init__(
-        self,
-        dim: int,
-        dim_head: int = 32,
-        window_size: int = 8,
-        dim_ff: int | None = None,
-        attn_drop_rate: float = 0.0,
-        ff_drop_rate: float = 0.0,
-        bias: bool = False,
-    ):
-        super().__init__()
-
-        if dim_ff is None:
-            dim_ff = dim * 4
-        heads = dim // dim_head
-
-        self.dim = dim
-        self.heads = heads
-        self.dim_head = dim_head
-        self.window_size = window_size
-        self.attn_drop_rate = attn_drop_rate
-        self.ff_drop_rate = ff_drop_rate
-        self.bias = bias
-
-        self.fused_dims = (dim, dim, dim_ff)
-        self.fused_attn_ff_proj = nn.Linear(dim, sum(self.fused_dims), bias=bias)
-        self.attn_gate = nn.Linear(dim, heads)  # bias always set to True
-        self.attn_out = nn.Linear(dim, dim, bias=bias)
-        self.ff_out = nn.Sequential(
-            StarReLU(),
-            nn.Dropout(ff_drop_rate),
-            nn.Linear(dim_ff, dim, bias=bias),
-        )
-
-    def _init_weights(self) -> None:
-        nn.init.constant_(self.attn_gate.bias, 0.5)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        x = rearrange(x, "b c (h p1) (w p2) -> b h w (p1 p2) c", p1=self.window_size, p2=self.window_size)
-        x, ps = pack([x], "* n d")
-
-        q, kv, x_ff = self.fused_attn_ff_proj(x).split(self.fused_dims, dim=-1)
-
-        q, kv = map(lambda t: rearrange(t, "b n (h d) -> b h n d", h=self.heads), (q, kv))
-        x_attn = F.scaled_dot_product_attention(q, kv, kv, dropout_p=self.attn_drop_rate if self.training else 0.0)
-        x_gate = self.attn_gate(x)  # shape = (b, n, h)
-        x_gate = rearrange(x_gate, "b n h -> b h n ()")
-        x_attn = x_attn * torch.sigmoid(x_gate)
-        x_attn = rearrange(x_attn, "b h n d -> b n (h d)")
-
-        x = self.attn_out(x_attn) + self.ff_out(x_ff)
-
-        (x,) = unpack(x, ps, "* n d")
-        x = rearrange(x, "b h w (p1 p2) c -> b c (h p1) (w p2)", p1=self.window_size, p2=self.window_size)
-        return x
 
 
 class FusedS2ViTBlock(nn.Module):
@@ -292,4 +231,4 @@ class S2ViT(nn.Module):
         return x
 
 
-__all__ = ["FusedGWAttentionFF", "S2ViT", "FusedS2ViTBlock", "ParallelS2ViTBlock", "S2ViTStage"]
+__all__ = ["S2ViT", "FusedS2ViTBlock", "ParallelS2ViTBlock", "S2ViTStage"]
